Updated readme
Factoy Escape

In this game you are a robot that should escape from the factory he's
been recluded,
but beware of acid waste and other dangers.

Good luck and have fun!

Controls:

- A: Move left
- D: Move right
- SPACE: Jump
- LSHIFT: Slide
- E: Melee

- KP -: Volume down
- KP +: Volume up

- O: Enter photo mode
- Arrow Up: Move the camera up
- Arrow Down: Move the camera down
- Arrow Right: Move the camera to the right
- Arrow Left: Move the camera to the left
- P: Take a screenshot, will be saved in the screenshot folder with the
date it was taken
- O: Exit the photo mode

- Escape: Exit the game

Debug Controls:

- F1: Start from the very first level
- F2: Start from the beginnin of the current level
- F5: Save the current state
- F6: Load previous state
- F9: Show/Hide logic layer
- F10: Activade/Deactivate player God Mode
- F11: Enable/Disable framerate cap to 30

GitHub:

https://github.com/OscarHernandezG/GameDevelopment/

- Authors

- Aleix: https://github.com/aleixgab
- Óscar: https://github.com/OscarHernandezG

MIT License

Copyright (c) 2017 Óscar Hernández and Aleix Gabarró

Permission is hereby granted, free of charge, to any person obtaining a
copy
of this software and associated documentation files (the "Software"), to
deal
in the Software without restriction, including without limitation the
rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or
sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included
in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
IN THE
SOFTWARE.